refactor : 메뉴 컨트롤러 클래스 분리 및 구글 컨벤션 적용
- 메뉴 컨트롤러 역할에 따라 클래스 분리 진행
- 구글 코드 컨벤션에 따라 코드 리펙토링
