+++ src/main/java/com/devcourse/web2_1_dashbunny_be/feature/admin/store/service/StoreManagementService.java
@@ -7,7 +7,7 @@ import com.devcourse.web2_1_dashbunny_be.domain.owner.StoreManagement;
 import com.devcourse.web2_1_dashbunny_be.domain.owner.role.StoreStatus;
 import com.devcourse.web2_1_dashbunny_be.feature.admin.store.dto.StoreCreateRequestDTO;
 import com.devcourse.web2_1_dashbunny_be.feature.admin.store.repository.StoreApplicationRepository;
-import com.devcourse.web2_1_dashbunny_be.feature.admin.store.repository.StoreManagementRepository;
+import com.devcourse.web2_1_dashbunny_be.feature.owner.store.repository.StoreManagementRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -43,9 +43,6 @@ public class StoreManagementService {
         storeManagement.setContactNumber(storeCreateRequestDTO.getContactNumber());
         storeManagement.setDescription(storeCreateRequestDTO.getDescription());
         storeManagement.setAddress(storeCreateRequestDTO.getAddress());
-        storeManagement.setCategory1(storeCreateRequestDTO.getCategory1());
-        storeManagement.setCategory2(storeCreateRequestDTO.getCategory2());
-        storeManagement.setCategory3(storeCreateRequestDTO.getCategory3());
         storeManagement.setStoreRegistrationDocs(storeCreateRequestDTO.getStoreRegistrationDocs());
         storeManagement.setStoreBannerImage(storeCreateRequestDTO.getStoreBannerImage());
         storeManagement.setStoreStatus(StoreStatus.PENDING); // 상태를 재등록 신청 중으로 변경
