+++ src/main/java/com/devcourse/web2_1_dashbunny_be/feature/owner/store/service/StoreService.java
@@ -3,10 +3,20 @@ package com.devcourse.web2_1_dashbunny_be.feature.owner.store.service;
 
 import com.devcourse.web2_1_dashbunny_be.feature.owner.dto.store.*;
 
+/**
+ * 가게 관리를 위한 service interface.
+ */
 public interface StoreService {
-    BasicInfoListResponseDto findBasicInfo(String storeId);
 
-    void updateBasicInfo(UpdateBasicInfoRequestDto updateBasicInfo);
+  /**
+   * 기본 정보를 반환 api service.
+   */
+  BasicInfoListResponseDto findBasicInfo(String storeId);
+
+  /**
+   * 기본 정보 수정을 위한 api service.
+   */
+  void updateBasicInfo(UpdateBasicInfoRequestDto updateBasicInfo);
 
     OperationInfoListResponseDto findOperationInfo(String storeId);
